pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["all"], version = "^0.110.0"}
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.6.1"}
python-dotenv = "^1.0.1"
celery = "^5.3.6"
redis = "^5.0.1"
tenacity = "^8.2.3"
httpx = "^0.27.0"
ruff = "^0.2.1"
faker = "^23.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "^42.0.5"
rich = "^13.7.0"
asyncio-throttle = "^1.1.0"
aiofiles = "^23.2.1"
python-multipart = "^0.0.9"
minio = "^7.1.17"
pyarrow = "^15.0.0"
pandas = "^2.2.0"
lxml = "^4.9.4"
beautifulsoup4 = "^4.12.3"
openpyxl = "^3.1.2"
oracledb = "^2.1.0"
pymysql = "^1.1.0"
python-tds = "^1.10.0"
snowflake-sqlalchemy = "^1.4.7"
google-cloud-bigquery = "^3.18.0"
google-cloud-storage = "^2.15.0"
snowflake-connector-python = "^3.6.0"
celery-beat-sqlalchemy = "^0.4.0" # Added for persistent scheduling


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pre-commit = "^3.6.0"
respx = "^0.20.1"
docker = "^7.0.0"
mypy = "^1.8.0"
types-redis = "^4.6.0.20240106"
types-python-jose = "^3.3.4.20240125"
types-passlib = "^1.7.7.20240106"
types-psycopg2 = "^2.9.21.20240128"
types-requests = "^2.31.0.20240125"
types-sqlalchemy = "^2.0.0.20240201"
ruff = "^0.2.1"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Unused import in __init__.py is often fine.
"tests/manual_scripts/*" = ["T201"] # Allow print statements in manual scripts.

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
line-ending = "auto"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
warn_unreachable = true
plugins = [
    "pydantic.mypy"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
